sparrow: insar display improvements
- color handling and general setup
- float, int -> float64, int64 (numpy v1.24)
- right display of NaN values
- adopt CPTHandler for Kite Scene display purposes
- update internal functions

+++ src/geometry.py
@@ -14,14 +14,14 @@ r2d = 1.0 / d2r
 
 
 def arr_vertices(obj):
-    a = num.array(obj, dtype=num.float)
+    a = num.array(obj, dtype=num.float64)
     assert len(a.shape) == 2
     assert a.shape[1] == 3
     return a
 
 
 def arr_faces(obj):
-    a = num.array(obj, dtype=num.int)
+    a = num.array(obj, dtype=num.int64)
     assert len(a.shape) == 2
     return a
 
@@ -50,7 +50,7 @@ def rtp2xyz(rtp):
     r = rtp[:, 0]
     theta = rtp[:, 1]
     phi = rtp[:, 2]
-    vecs = num.empty(rtp.shape, dtype=num.float)
+    vecs = num.empty(rtp.shape, dtype=num.float64)
     vecs[:, 0] = r*num.sin(theta)*num.cos(phi)
     vecs[:, 1] = r*num.sin(theta)*num.sin(phi)
     vecs[:, 2] = -r*num.cos(theta)
@@ -59,7 +59,7 @@ def rtp2xyz(rtp):
 
 def xyz2rtp(xyz):
     x, y, z = xyz[:, 0], xyz[:, 1], xyz[:, 2]
-    vecs = num.empty(xyz.shape, dtype=num.float)
+    vecs = num.empty(xyz.shape, dtype=num.float64)
     vecs[:, 0] = num.sqrt(x**2+y**2+z**2)
     vecs[:, 1] = num.arctan2(num.sqrt(x**2+y**2), z)
     vecs[:, 2] = num.arctan2(y, x)
@@ -114,7 +114,7 @@ def topo_to_faces(nlat, nlon):
     if k not in g_topo_faces:
 
         nfaces = 2 * (nlat - 1) * (nlon - 1)
-        faces = num.empty((nfaces, 3), dtype=num.int)
+        faces = num.empty((nfaces, 3), dtype=num.int64)
         ilon = num.arange(nlon - 1)
         for ilat in range(nlat - 1):
             ibeg = ilat*(nlon-1) * 2
@@ -139,7 +139,7 @@ def topo_to_faces_quad(nlat, nlon):
     if k not in g_topo_faces_quad:
 
         nfaces = (nlat - 1) * (nlon - 1)
-        faces = num.empty((nfaces, 4), dtype=num.int)
+        faces = num.empty((nfaces, 4), dtype=num.int64)
         ilon = num.arange(nlon - 1)
         for ilat in range(nlat - 1):
             ibeg = ilat*(nlon-1)
@@ -178,11 +178,11 @@ def refine_triangles(vertices, faces):
     ne = fedges.shape[0] // 2
     edges = fedges[sortorder[::2], :]
 
-    vertices_new = num.zeros((nv+ne, 3), dtype=num.float)
+    vertices_new = num.zeros((nv+ne, 3), dtype=num.float64)
     vertices_new[:nv] = vertices
     vertices_new[nv:] = 0.5 * (vertices[edges[:, 0]] + vertices[edges[:, 1]])
 
-    faces_new = num.zeros((nf*4, 3), dtype=num.int)
+    faces_new = num.zeros((nf*4, 3), dtype=num.int64)
 
     vind = nv + sortorder.argsort()/2
